# Api.yaml
openapi: 3.0.0
info:
  version: 1.0.0
  title: Express Gateway API
  description: Provides centralized access endpoints for VUE UI application to access all other available services.
tags:
  - name: Security services
    description: Provides secure login and user credential translation to jwt token either using remote ANAT/LDAP service or local mock implementation
  - name: Gateway services
    description: Provides secured access to backend services, validating JWT token. Expected operations to be supported are GET (WMJ, MMS, CoCo), PUT (MMS)

paths:
  /login:
    post:
      tags:
        - Security services
      description: Authenticate user and return roles and jwt token
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Security artifact successfuly generated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        "401":
          description: Forbidden due wrong or missing privilegies
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "403":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "404":
          description: Api endpoint not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "5XX":
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  
  /login/verify:
    get:
      tags:
        - Security services
      description: Verify token of the authenticated user
      parameters:
        - $ref: "#/components/parameters/ibm-sec-token"
      
      responses:
        "200":
          description: Security artifact successfuly verified
          
        "401":
          description: Forbidden due wrong or missing privilegies
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "403":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "404":
          description: Api endpoint not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "5XX":
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
                
  /gateway/XXX:
    get:
      tags:
        - Gateway services
      description: Provides GET routing. Currently only MMS and WMJ are supported.<br>Service is identified by XXX code (currently supported wmj, mms). <br>**IT IS NOT POSSIBLE TO TEST** the api due to inability to express generic nature of query part of URL path in OAS3 format.<br>Example of supported call from Postman or provided testcases (/tests/gateway.test.js) <br>* **/gateway/mms?kmat=hrasko**

      parameters:
        - $ref: "#/components/parameters/ibm-sec-token"

      responses:
        "200":
          description: Security artifact successfuly verified
          
        "401":
          description: Forbidden due wrong or missing privilegies
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "403":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "404":
          description: Api endpoint not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "5XX":
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

components:
  
  parameters:
   
    ibm-sec-token:
      name: ibm-sec-token
      in: header
      required: true
      description: JWT security token for mutual trust between apps
      schema:
        type: string


    

  schemas:

    Error:
      type: object
      description: Chybové hlášení
      properties:
        status:
          type: string
        statusCase:
          type: string
        msg:
          type: string
      required:
        - status
        - statusCase
        - msg

    LoginRequest:
      type: object
      description: Authentication request
      properties:
        name:
          type: string
        password:
          type: string
      required:
        - name
        - password
      
    LoginResponse:
      type: object
      description: Authentication response
      properties:
        name:
          type: string
        id:
          type: string
        email:
          type: string
        role:
          type: array
          items:
            type: string
        token:
          type: string
        status:
          type: string
        statusCase:
          type: string
        msg:
          type: string
          
      required:
        - name
        - id
        - email
        - role
        - token
        - status
        - statusCase
        - msg

    GenericRequest:
      type: object
      description: Authentication request
      

    GenericResponse:
      type: object
      description: Authentication request
      

    TokenRequest:
      type: object
      description: JWT token test request for security admin/participant and for admin/business app (notarius) scenarios. No dummy jwt tokens are accepted
      properties:
        user_id:
          type: string
        company_id:
          type: string
        iss:
          type: string
        
      required:
        - user_id
        - company_id
        - iss
    
    TokenResponse:
      type: object
      description: JWT token test response for use in other API calls
      properties:
        token:
          type: string

    

      



 